requirements.txt
requests>=2.31.0
pandas>=2.0.0
scikit-learn>=1.3.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
joblib>=1.3.0
